+++ src/fs/clean.ts
